refactor: split transform.ts god module into focused modules
- Split 393-line transform.ts into 7 focused modules under src/transform/
- Fix CaseConverter singleton (was creating new instance per call)
- Create shared TransformOptions type

New structure:
  src/transform/
    index.ts      - orchestration + re-exports
    types.ts      - TransformOptions
    utils.ts      - toPascalCase, singularize
    type-mapper.ts - mapPostgresType
    filter.ts     - filterTables
    enum.ts       - transformEnum
    table.ts      - transformTable, transformColumn, createDBInterface

Closes #3
