+++ src/transform/types.ts
@@ -1,0 +1,5 @@
+export type TransformOptions = {
+  camelCase?: boolean;
+  includePattern?: string[];
+  excludePattern?: string[];
+};
