+++ src/utils/case-converter.ts
@@ -16,6 +16,8 @@ class CaseConverter extends CamelCasePlugin {
  * toCamelCase('foo_bar') // => 'fooBar'
  * toCamelCase('created_at') // => 'createdAt'
  */
+const caseConverter = new CaseConverter();
+
 export function toCamelCase(str: string): string {
-  return new CaseConverter().toCamelCase(str);
+  return caseConverter.toCamelCase(str);
 }
